Refactor modules into shared pipeline stages
