+++ module_singleclickicore.py
@@ -1,17 +1,5 @@
-import logging
-import os
-
-from module_imagedeid_pacs import imagedeid_pacs
-from module_textdeid import textdeid
-from module_image_export import image_export
-from utils import (
-    PacsConfiguration,
-    RunDirs,
-    SingleClickResult,
-    Spreadsheet,
-    setup_run_directories,
-    configure_run_logging,
-)
+from pipeline import SingleClickIcorePipeline
+from utils import PacsConfiguration, RunDirs, SingleClickResult, Spreadsheet
 
 
 def singleclickicore(
@@ -44,62 +32,15 @@ def singleclickicore(
     deferred_delivery_timeout: int = 172800,
     deid_engine: str = "ctp",
 ) -> SingleClickResult:
-    """
-    Combined module that performs:
-    1. Image deidentification from PACS (HIPAA Safe Harbor enforced)
-    2. Text deidentification on input Excel file
-    3. Export of all output to Azure Blob Storage (optional)
-
-    Args:
-        pacs_list: List of PacsConfiguration objects
-        query_spreadsheet: Spreadsheet object for PACS querying
-        application_aet: Application AE Title for DICOM communication
-        sas_url: SAS URL for Azure container
-        project_name: Project name (used as folder prefix in blob storage)
-        output_dir: Output directory for deidentified files
-        input_file: Path to input Excel file (for text deid)
-        appdata_dir: Application data directory for logs/metadata
-        filter_script: CTP filter script (already combined with HIPAA quarantine filter by config builder)
-        date_window_days: Date window for PACS queries
-        anonymizer_script: Pre-generated HIPAA Safe Harbor anonymizer script (from config builder)
-        deid_pixels: Enable pixel de-identification (always True for HIPAA)
-        lookup_table: CTP lookup table
-        apply_default_filter_script: Whether to apply default CTP filter
-        mapping_file_path: Path to mapping file for value substitution
-        to_keep_list: List of terms to preserve during text deid
-        to_remove_list: List of terms to redact during text deid
-        columns_to_drop: List of columns to drop from Excel
-        columns_to_deid: List of columns to deidentify in Excel
-        debug: Enable debug logging
-        run_dirs: Run directories dictionary
-        skip_export: Whether to skip Azure export (default: False)
-
-    Returns:
-        dict: Combined results from all steps
-
-    HIPAA Safe Harbor Settings (Automatically Enforced):
-        - Anonymizer: HIPAA Safe Harbor script (generated by config builder)
-        - Pixel De-identification: ENABLED (always True)
-    """
-    if run_dirs is None:
-        run_dirs = setup_run_directories()
-
-    log_level = logging.DEBUG if debug else logging.INFO
-    configure_run_logging(run_dirs["run_log_path"], log_level)
-    logging.info("Running singleclickicore")
-
-    if appdata_dir is None:
-        appdata_dir = run_dirs["appdata_dir"]
-
-    os.makedirs(appdata_dir, exist_ok=True)
-    os.makedirs(output_dir, exist_ok=True)
-
-    # Step 1: Image deid from PACS
-    logging.info("=" * 80)
-    logging.info("STEP 1: Image Deidentification from PACS")
-    logging.info("=" * 80)
-
-    deid_result = imagedeid_pacs(
+    return SingleClickIcorePipeline(
+        sas_url=sas_url,
+        project_name=project_name,
+        input_file=input_file,
+        to_keep_list=to_keep_list,
+        to_remove_list=to_remove_list,
+        columns_to_drop=columns_to_drop,
+        columns_to_deid=columns_to_deid,
+        skip_export=skip_export,
         pacs_list=pacs_list,
         query_spreadsheet=query_spreadsheet,
         application_aet=application_aet,
@@ -119,61 +60,4 @@ def singleclickicore(
         storescp_port=storescp_port,
         deferred_delivery=deferred_delivery,
         deferred_delivery_timeout=deferred_delivery_timeout,
-        deid_engine=deid_engine,
-    )
-
-    # Step 2: Text deid on input Excel
-    logging.info("=" * 80)
-    logging.info("STEP 2: Text Deidentification on Input File")
-    logging.info("=" * 80)
-
-    text_result = textdeid(
-        input_file=input_file,
-        output_dir=output_dir,
-        to_keep_list=to_keep_list,
-        to_remove_list=to_remove_list,
-        columns_to_drop=columns_to_drop,
-        columns_to_deid=columns_to_deid,
-        debug=debug,
-        run_dirs=run_dirs,
-    )
-
-    # Step 3: Export to Azure (only if we have content to export and not skipped)
-    if not skip_export:
-        logging.info("=" * 80)
-        logging.info("STEP 3: Export to Azure Blob Storage")
-        logging.info("=" * 80)
-
-        if sas_url is None:
-            raise ValueError("sas_url is required when skip_export is False")
-
-        if deid_result["num_images_saved"] > 0 or text_result["num_rows_processed"] > 0:
-            image_export(
-                input_dir=output_dir,
-                sas_url=sas_url,
-                project_name=project_name,
-                appdata_dir=appdata_dir,
-                debug=debug,
-                run_dirs=run_dirs,
-            )
-        else:
-            logging.info("No content to export - skipping Azure upload")
-    else:
-        logging.info("=" * 80)
-        logging.info("STEP 3: Export to Azure Blob Storage - SKIPPED")
-        logging.info("=" * 80)
-
-    logging.info("=" * 80)
-    logging.info("singleclickicore complete")
-    logging.info(f"Deidentified files preserved at: {output_dir}")
-    logging.info("=" * 80)
-
-    return {
-        "num_studies_found": deid_result["num_studies_found"],
-        "num_images_exported": deid_result["num_images_saved"],
-        "num_images_quarantined": deid_result["num_images_quarantined"],
-        "failed_query_indices": deid_result["failed_query_indices"],
-        "num_rows_processed": text_result["num_rows_processed"],
-        "output_file": text_result["output_file"],
-        "export_performed": not skip_export,
-    }
+    ).run()
